Create faq.gui.js
When Galacta Knight is defeated in Meta Knightmare Ultra and The True Arena, he flickers white and flies off into space. In Meta Knightmare Ultra, a short cutscene plays, where Galacta Knight is shown flying unstably in space, then rapidly appearing in multiple parts of the screen, as if he was exploding. He then appears in the middle of the screen where he is engulfed in a bright, white flash. Galacta Knight's wings are torn off his body, while his sword and shield is shown leaving him as well. The disarmed knight then vanishes in the flash, while Meta Knight returns to Planet Popstar. In The True Arena, this cutscene is not shown, as Kirby now has to fight Marx Soul in the final round