Added new regex command (split by ∆)
First off, let’s start with the most iconic lore characters: the ancients.

The ancients, as their name implies, are people from long, long ago that are no longer around (or so we think). The ancients are iconic in the lore for being affiliated with almost everything. Let’s take a look at their history, shall we? The only planet confirmed to have been inhabited by the ancients is Halcandra, which is located in another dimension. Now, that’s not to say they didn’t inhabit Kirby’s dimension, as several of the ancients’ creations can be found there. Regardless, the ancients were highly advanced in science, as they created many legendary pieces of technology. According to Magolor, the Lor Starcutter, a ship that can traverse through space and time, was crafted by the ancients using an incredible power. He also states that they created “clockwork stars that soar the cosmos” using the same power. Many have presumed these stars to be Novas, clockwork cat faces that can grant wishes. This was seemingly proven in Kirby: Planet Robobot, where the final boss of the game, Star Dream, fused with the Access Ark, revealing it to actually be a Nova. According to Haltmann, his company analyzed the most advanced civilizations in the known universe and used blueprints they found in order to reactivate Star Dream. The Lor and the Novas were two examples of their skill in science, but they were also skilled in magic as well. Also according to Magolor, the ancients crafted “mysterious items that can bring dreams to life.” These are presumed to be the Star Rods, which power the Fountains of Dreams. Each fountain is responsible for giving every inhabitant in their respective planet good dreams. They also created the Master Crown, a crown that grants its wearer limitless power. However, the crown itself appears to be sentient, as it takes control of Magolor’s body and transforms him into a manifestation of itself. In this form, Magolor notably has an eye in his mouth. Many have thought the crown was Dark Matter because of this. While almost everyone else thought that wasn’t the case, those who said it was just might have been correct. We’ll get to that later. Anyway, those two are examples of the ancients’ skill in magic. Slightly off-topic, but since the ancients lived on Halcandra, it is likely that they made Egg Engines, and in turn, Metal General, which the Haltmann Works Company probably used the blueprints of in order to make the Security Force. The ancients also entrusted Landia, a four-headed dragon, in guarding the legendary treasures they made on that planet.


However, that’s not even close to all the information about them...