Now Split by ~
As Kirby Star Allies is the most lore-heavy game as of now, they wouldn’t dare to not include the ancients. As we all know, Hyness, the main antagonist of the game, dropped a lore bomb in the form of having a stroke and spouting it out at light speed. Given that the English localization butchered this lore bomb, we’ll take a look at the original Japanese version. According to Hyness, the people of his clan were powerful wielders of magic, and were friends with people who used the power of science. Hmm, sounds familiar. The people of his clan were also worshippers of a matter most dark. Anyway, one day, their science-using friends turned on them in fear of their power and banished them to the edge of the galaxy. As if that wasn’t enough, they then subsequently erased all but a sliver of their very existence from history itself. So yeah, most of them are dead. I think Hyness is like, a descendant from this sliver, but I have no idea since it’s totally unclear whether he’s the last survivor or a descendant. Anyway, he also states that his clan was responsible for stopping a galactic crisis of sorts. Why am I telling you about Hyness’s clan? Well, it’s been theorized that Hyness’s clan and these science-users WERE the ancients, and it kind of makes sense. Think of it this way, the science-users build the base of the creation (specifically the Lor and the Novas), and Hyness’s clan imbues them with the “incredible power,” in other words, magic. That would explain how the Lor can traverse space and time and how the Novas can grant wishes. The clan most likely also made the Star Rods and the FoD, which have more to do with magic than science. Now, moving on to the Master Crown. It is theorized that it has something to do with Dark Matter, right? Given that the clan were worshippers of a matter most dark, it could be possible that they made the crown using Dark Matter. That would explain why Void Termina, the final boss of Star Allies and the reincarnation of the implied ancestor of Dark Matter, uses a replica of it in his boss fight. Speaking of Void Termina, his normal self, simply known as Void, is reportedly spoken about in several books that were written by the ancients after witnessing his starlike radiance. One of these books is likely the book of legend, which basically called Void a god. This book is presumed to be written by Hyness’s clan, due to him saying that everyone else thought it was a fairytale. And so, that is all the information surrounding the ancients so far. This is...really long. Anyway, next time, we’ll be talking about our favorite one-eyed family that corrupts planets, as well as their baby-faced presumed ancestor.